+++ lib/rugged/diff.rb
@@ -7,9 +7,6 @@ module Rugged
     include Enumerable
     alias each each_patch
 
-    attr_reader :owner
-    alias tree owner
-
     def patches
       each_patch.to_a
     end