+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Remove `Rugged::Diff#tree` and `Rugged::Diff#owner`.
+
+    Fixes #389.
+
+    *Arthur Schreiber*
+
 *   Add `#additions` and `#deletions` to `Rugged::Patch`.
 
     *Mindaugas Mozūras*